requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
idna==2.10
Pillow==8.3.0
psycopg2==2.9.1
pycryptodome==3.10.1
PyJWT==2.1.0
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
requests==2.25.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
webdriver-manager==3.4.2
